+++ packages/assignment/src/provider/UserProvider.tsx
@@ -1,0 +1,34 @@
+import React, { createContext, useContext, useState } from 'react';
+
+interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+const UserContext = createContext<UserContextType | undefined>(undefined);
+
+export const useUserClient = (): UserContextType => {
+  const context = useContext(UserContext);
+
+  if (!context) {
+    throw new Error('UserProvider error');
+  }
+
+  return context;
+};
+
+export const UserProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+  const [user, setUser] = useState<User | null>(null);
+
+  const login = (email: string) => setUser({ id: 1, name: '홍길동', email });
+  const logout = () => setUser(null);
+
+  return <UserContext.Provider value={{ user, login, logout }}>{children}</UserContext.Provider>;
+};
